Align schema with the live prod dump: CHECKs, uniques, types, nullability
Diffed the port against the real Supabase schema dump (every CHECK, FK,
UNIQUE and default) and fixed the gaps in migration AlignWithProdSchema:

- 3 functional bugs: messages.type now allows workout_plan/nutrition_plan
  (was rejecting real plan messages); barcode_products.source includes
  'manual'; messaging writes notifications.type='message' (CHECK allows
  message|plan only)
- Open question resolved: notifications.coach_id -> profiles(id) with FK;
  notifications.plan_id has no FK in prod (extra FK removed)
- barcode_scan_history.barcode is nullable with no FK (aligned)
- UNIQUE indexes: onboarding/user_goals/user_active_program/coaches
  .user_id, stripe_customers.user_id + stripe_customer_id
- ~22 CHECK constraints added with prod-confirmed value lists
- Type fixes: weekly_activity pcts -> INT (view AVG now CASTs),
  daily_activity heart_rate/exercise_minutes -> numeric
- ~25 columns aligned to prod NOT NULL/defaults (incl. daily_summary
  zero defaults, coach_onboarding required fields, JSON arrays)
- Flagged prod contradiction: subscription_plans.coach_id FK->coaches.id
  vs RLS auth.uid() (port follows RLS; decision query in checklist)
- 127/127 tests, verify.ps1 23/23

+++ src/CoreGym.Domain/Entities/DailyActivity.cs
@@ -13,9 +13,10 @@ public class DailyActivity
 
     public decimal ActiveCaloriesBurned { get; set; }
 
-    public int? HeartRateAvg { get; set; }
+    /// <summary>Live prod dump: numeric, not integer.</summary>
+    public decimal? HeartRateAvg { get; set; }
 
-    public int? ExerciseMinutes { get; set; }
+    public decimal? ExerciseMinutes { get; set; }
 
     public string? Source { get; set; }
